+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 
 * Require `webdriver` v4
 * Require `watir` v7
+* `Selenium::WebDriver::Error::ElementNotVisibleError` replaced with
+  `Selenium::WebDriver::Error::ElementNotInteractableError`
 * Remove `ruby-2.5` from CI since it's EOLed
 
 ## 0.22.1 (2021-12-17)