+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 * Require `watir` v7
 * `Selenium::WebDriver::Error::ElementNotVisibleError` replaced with
   `Selenium::WebDriver::Error::ElementNotInteractableError`
+* `Selenium::WebDriver::Error::ElementNotInteractableError` replaced with
+  `Selenium::WebDriver::Error::InvalidElementStateError`
 * Remove `ruby-2.5` from CI since it's EOLed
 
 ## 0.22.1 (2021-12-17)